requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.18.42
botocore==1.21.42
certifi==2021.5.30
cffi==1.14.6
Django==3.2.7
django-cors-headers==3.7.0
jmespath==0.10.0
mysqlclient==2.0.3
pycparser==2.20
PyMySQL==1.0.2
bcrypt==3.2.0
pyjwt==2.1.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
requests==2.26.0
